+++ src/main/java/cef/financial/domain/dto/SimulationHistoryResponseDTO.java
@@ -1,13 +1,14 @@
 package cef.financial.domain.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.persistence.criteria.CriteriaBuilder;
 
 import java.time.OffsetDateTime;
 
 public class SimulationHistoryResponseDTO {
 
-    public Long id;
-    public Long clienteId;
+    public Long   id;
+    public Integer clienteId;
     public String produto;
     public double valorInvestido;
     public double valorFinal;
